feat(moderation): create overview and guest pages - update components - remove redundant files - update routes and subroutes
