+++ ui/apps/moderation/src/components/app-routes.tsx
@@ -5,9 +5,9 @@ import { useCheckModerator, useGetLogin } from '@akashaorg/ui-awf-hooks';
 import { RootComponentProps } from '@akashaorg/typings/ui';
 
 import TransparencyLog from './transparency-log';
-import { Dashboard, GuestPage, IntroPage, ValuePage } from '../pages';
+import { Dashboard, Overview, ValuePage } from '../pages';
 
-import routes, { GUEST, HISTORY, HOME, MODERATION_VALUE, UNAUTHENTICATED } from '../routes';
+import routes, { DASHBOARD, HISTORY, HOME, MODERATION_VALUE } from '../routes';
 
 const { Box } = DS;
 
@@ -25,23 +25,12 @@ const AppRoutes: React.FC<RootComponentProps> = props => {
     <Box>
       <Router basename={props.baseRouteName}>
         <Routes>
-          <Route
-            path={routes[UNAUTHENTICATED]}
-            element={
-              <IntroPage {...props} user={loginQuery.data?.pubKey} isAuthorised={isAuthorised} />
-            }
-          />
+          <Route path={routes[HOME]} element={<Overview {...props} />} />
 
           <Route path={routes[MODERATION_VALUE]} element={<ValuePage {...props} />} />
 
           <Route
-            path={routes[GUEST]}
-            element={
-              <GuestPage {...props} user={loginQuery.data?.pubKey} isAuthorised={isAuthorised} />
-            }
-          />
-          <Route
-            path={routes[HOME]}
+            path={routes[DASHBOARD]}
             element={
               <Dashboard
                 {...props}
@@ -51,6 +40,7 @@ const AppRoutes: React.FC<RootComponentProps> = props => {
               />
             }
           />
+
           <Route
             path={routes[HISTORY]}
             element={
@@ -60,6 +50,7 @@ const AppRoutes: React.FC<RootComponentProps> = props => {
               />
             }
           />
+
           <Route path="/" element={<Navigate to={routes[HOME]} replace />} />
         </Routes>
       </Router>
