+++ ui/apps/moderation/src/interfaces/index.ts
@@ -1,2 +1,1 @@
 export * from './content';
-export * from './shared';
