--- ui/apps/moderation/src/interfaces/shared.ts
@@ -1,8 +1,0 @@
-import { RootComponentProps } from '@akashaorg/typings/ui';
-
-export interface ISharedModerationProps extends RootComponentProps {
-  slotId?: string;
-  user: string | null;
-  isAuthorised: boolean;
-  handleCTAClick?: () => void;
-}
