+++ ui/apps/moderation/src/pages/index.tsx
@@ -1,6 +1,5 @@
 import Dashboard from './dashboard';
-import IntroPage from './intro';
-import GuestPage from './guest';
+import Overview from './overview';
 import ValuePage from './moderation-value';
 
-export { Dashboard, IntroPage, GuestPage, ValuePage };
+export { Dashboard, Overview, ValuePage };
